Convert some files to typescript
