+++ tsconfig.json
@@ -3,15 +3,15 @@
 		"js/**/*"
 	],
   "compilerOptions": {
-    "target": "ES6",
+    "target": "ES2020",
     "useDefineForClassFields": true,
-    "module": "ES6",
+    "module": "ES2020",
     "moduleResolution": "Node",
     "jsx": "preserve",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "esModuleInterop": true,
-    "lib": ["ES6", "DOM"],
+    "lib": ["ES2020", "DOM"],
     "skipLibCheck": true,
     "noEmit": true,
     "allowJs": true
